tools/changelogs/requirements.txt
# Changelog tools requirements
PyGithub==1.55
PyYaml==6.0
